sys-assert: apply registers for 64bit architecture
- The registers of 64 bit architecture are little bit different
  with 32 bit architecture. Thus the operations to print
  register information are separated according to the architectures.

+++ src/sys-assert/sys-assert.c
@@ -103,7 +103,7 @@ static int trace_symbols(void *const *array, int size, struct addr_node *start, 
 	Elf32_Sym *symtab_entry;
 	int i, cnt, file, ret;
 	char *fname;
-	unsigned int addr, start_addr, offset_addr;
+	unsigned long addr, start_addr, offset_addr;
 	unsigned int strtab_index = 0;
 	unsigned int symtab_index = 0;
 	int num_st = 0;
@@ -120,8 +120,8 @@ static int trace_symbols(void *const *array, int size, struct addr_node *start, 
 					cnt, array[cnt], dlerror());
 			continue;
 		}
-		start_addr = (unsigned int)get_start_addr(array[cnt], start);
-		addr = (unsigned int)array[cnt];
+		start_addr = (unsigned long)get_start_addr(array[cnt], start);
+		addr = (unsigned long)array[cnt];
 		/* because of launchpad,
 		 * return value of dladdr when find executable is wrong.
 		 * so fix dli_fname here */
@@ -391,8 +391,8 @@ static void print_node_to_file(struct addr_node *start, int fd)
 		} else {
 			fprintf_fd(fd,
 					"%08x %08x %s %s\n",
-					(unsigned int)t_node->startaddr,
-					(unsigned int)t_node->endaddr,
+					(unsigned long)t_node->startaddr,
+					(unsigned long)t_node->endaddr,
 					t_node->perm, t_node->fpath);
 			t_node = t_node->next;
 		}
